Fix signedness warning in tests/memchunks.c
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from tests/memchunks.c:35:
tests/memchunks.c: In function ‘old_mem_chunk_new’:
tests/memchunks.c:140:35: error: comparison of integer expressions of different signedness: ‘gulong’ {aka ‘long unsigned int’} and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
   g_return_val_if_fail (area_size >= atom_size, NULL);
                                   ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
tests/memchunks.c:140:3: note: in expansion of macro ‘g_return_val_if_fail’
   g_return_val_if_fail (area_size >= atom_size, NULL);
   ^~~~~~~~~~~~~~~~~~~~

+++ tests/memchunks.c
@@ -137,7 +137,7 @@ old_mem_chunk_new (const gchar  *name,
   gulong rarea_size;
   
   g_return_val_if_fail (atom_size > 0, NULL);
-  g_return_val_if_fail (area_size >= atom_size, NULL);
+  g_return_val_if_fail (area_size >= (gulong) atom_size, NULL);
   
   ENTER_MEM_CHUNK_ROUTINE ();
